Move ReferenceListerService to platform/base
