+++ platform/base/core/src/main/java/com/peregrine/reference/Reference.java
@@ -1,4 +1,4 @@
-package com.peregrine.replication;
+package com.peregrine.reference;
 
 /*-
  * #%L
