+++ platform/base/core/src/main/java/com/peregrine/replication/impl/DistributionReplicationService.java
@@ -28,7 +28,7 @@ package com.peregrine.replication.impl;
 import com.peregrine.replication.ReplicationServiceBase;
 import com.peregrine.commons.util.PerConstants;
 import com.peregrine.commons.util.PerUtil;
-import com.peregrine.replication.ReferenceLister;
+import com.peregrine.reference.ReferenceLister;
 import com.peregrine.replication.Replication;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.sling.api.resource.Resource;
